Calculate correct stack level for deprecation warnings to show them to the user.
Generating deprecation warnings with the correct stacklevel is tricky to get right. Up to now, users have not been seeing these deprecation warnigns because they had incorrect stacklevels. See https://github.com/python/cpython/issues/68493 and https://github.com/python/cpython/issues/67998 for good discussions, as well as https://github.com/ipython/ipython/issues/8478#issuecomment-106119371 and https://github.com/ipython/ipython/pull/8480#issuecomment-106107638 for more context.

We see these issues directly. Normally stacklevel=2 might be enough to target the caller of a function. However, in our deprecation warning in the __init__ method, each level of subclasses adds one stack frame, so depending on the subclass, the appropriate stacklevel is different. Thus this PR implements a trick from a discussion in Python to calculate the first stack frame that is external to the ipywidgets library, and use that as the target stack frame for the deprecation warning.